Project/AI/requirements.txt
flask
pandas
numpy
scikit-learn
huggingface_hub
torch
transformers
requests
gradio
python-dotenv
